requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
